+++ cmd/atcoderHelper/ach/version_test.go
@@ -1,4 +1,4 @@
-package cmd
+package ach
 
 import (
 	"bytes"
