--- cmd/atcoderHelper/cmd/root.go
@@ -1,29 +1,0 @@
-package cmd
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/spf13/cobra"
-)
-
-// rootCmd represents the base command when called without any subcommands
-var rootCmd = &cobra.Command{
-	Use:   "ach",
-	Short: "ach は Atcoder出場に際する定形作業を自動化します",
-	Long:  `ach は Atcoder出場に際する定形作業を自動化します。`,
-}
-
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
-}
-
-func init() {
-	cobra.OnInitialize(initConfig)
-}
-
-func initConfig() {
-}
