+++ cmd/gendocs/main.go
@@ -1,17 +1,17 @@
 package main
 
 import (
-	"fmt"
-	"os"
+	"log"
 
+	"github.com/spf13/cobra/doc"
 	"github.com/yuchiki/atcoderHelper/cmd/atcoderHelper/ach"
 )
 
 func main() {
 	cmd := ach.NewAchCmd()
 
-	if err := cmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+	err := doc.GenMarkdownTree(cmd, "docs/cmd")
+	if err != nil {
+		log.Fatal(err)
 	}
 }
