+++ docs/cmd/ach.md
@@ -1,3 +1,19 @@
-# ach
+## ach
 
-実装中...
+ach は Atcoder出場に際する定形作業を自動化します
+
+### Synopsis
+
+ach は Atcoder出場に際する定形作業を自動化します。
+
+### Options
+
+```
+  -h, --help   help for ach
+```
+
+### SEE ALSO
+
+* [ach version](ach_version.md)	 - show version
+
+###### Auto generated by spf13/cobra on 4-Feb-2021
