+++ docs/cmd/ach_version.md
@@ -1,0 +1,23 @@
+## ach version
+
+show version
+
+### Synopsis
+
+show version.
+
+```
+ach version [flags]
+```
+
+### Options
+
+```
+  -h, --help   help for version
+```
+
+### SEE ALSO
+
+* [ach](ach.md)	 - ach は Atcoder出場に際する定形作業を自動化します
+
+###### Auto generated by spf13/cobra on 4-Feb-2021
